pyproject.toml
[tool.poetry]
name = "gai-course-project"
version = "0.1.0"
description = ""
authors = ["Ivan Golov <ivangolov@icloud.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
torch = ">=2.2.2"
numpy = ">=1.26.4"
jupyter = "^1.1.1"
diffusers = "^0.32.2"
datasets = "^3.2.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
pandas = "^2.2.3"
gradio = "^5.14.0"
fastapi = "^0.115.8"
transformers = "^4.48.2"
peft = "^0.14.0"
tqdm = "^4.67.1"
torchvision = "^0.21.0"
clearml = "^1.18.0"
pyyaml = "^6.0.2"
pillow = "^11.1.0"
streamlit = "^1.44.1"
pydantic = "^2.11.3"
uvicorn = "^0.34.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
accelerate = "^1.6.0"
huggingface-hub = "^0.30.2"
safetensors = "^0.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
